rename: 마이페이지 탭 컨텐츠 파일 위치 이동
