+++ src/components/mypage/contents/content/myTravel/CheckedTravel.tsx
