+++ src/components/mypage/contents/content/myTravel/PastTravel.tsx
