+++ src/mocks/data/travel/mypage/travelListMock.ts
@@ -20,7 +20,7 @@ const travelListMock = {
     },
     {
       travelId: 15,
-      isDomestic: true,
+      isDomestic: false,
       travelName:
         '제주 한라산 등반 후 미식 여행. 주로 한식&인스타 카페 방문 예정. 이후 일정 미정. 채팅으로 일정 정하기 가능!',
       travelLocation: '제주',
@@ -42,7 +42,7 @@ const travelListMock = {
   pastTravel: [
     {
       travelId: 17,
-      isDomestic: true,
+      isDomestic: false,
       travelName: '부여로 떠나는 다함께 시골투어',
       travelLocation: '충남 부여',
       maxParticipant: 6,
@@ -69,6 +69,18 @@ const travelListMock = {
       maxParticipant: 12,
       currentParticipant: 12,
       startDate: '12/06',
+    },
+  ],
+
+  mySelfTravel: [
+    {
+      travelId: 17,
+      isDomestic: false,
+      travelName: '부여로 떠나는 다함께 시골투어',
+      travelLocation: '충남 부여',
+      maxParticipant: 6,
+      currentParticipant: 1,
+      startDate: '12/03',
     },
   ],
 };
